server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Aryan Tamang <aryantmg628@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
celery = "^5.4.0"
redis = "^5.0.4"
django-debug-toolbar = "^4.3.0"
djangorestframework-simplejwt = "^5.3.1"
django-dotenv = "^1.4.2"
drf-nested-routers = "^0.94.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
